fix pg13/pg14 compilation -- Var.varno is u32 in those bindings
в биндингах pgrx для pg13/pg14 Var.varno имеет тип u32, а в pg15+ --
i32. topn.rs писался и тестировался на pg15, из-за чего cargo test
--all в ci падал на pg13/pg14 с E0308/E0277 (mismatched types,
can't compare u32 with i32).

нормализуем varno к i32 явным кастом в двух местах чтения
(score_sort_varno и walker-проверка required_varno): varno -- малый
положительный индекс rangetable, каст безопасен; на pg15 это no-op.

+++ src/walker/topn.rs
@@ -283,7 +283,9 @@ unsafe fn score_sort_varno(
         {
             return None;
         }
-        return Some((*var).varno);
+        // varno is u32 in pg13/pg14 bindings and i32 in pg15+; normalize to
+        // i32 (it's a small positive rangetable index either way)
+        return Some((*var).varno as i32);
     }
 
     None
@@ -427,7 +429,8 @@ unsafe extern "C" fn expr_scan_walker(node: *mut pg_sys::Node, context: void_mut
     } else if is_a(node, pg_sys::NodeTag::T_Var) {
         if let Some(required_varno) = context.required_varno {
             let var = node as *mut pg_sys::Var;
-            if (*var).varlevelsup == 0 && (*var).varno == required_varno {
+            // varno cast: u32 in pg13/pg14 bindings, i32 in pg15+
+            if (*var).varlevelsup == 0 && (*var).varno as i32 == required_varno {
                 context.var_count += 1;
             } else {
                 context.foreign_var_count += 1;
